# openapi.yaml
openapi: 3.0.3
info:
  title: OTP Auth API
  description: A backend service for OTP-based authentication and user management
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
paths:
  /api/auth/request-otp:
    post:
      summary: Request an OTP
      description: Generate and send OTP to phone number (printed to console)
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: "+15551234567"
                  description: Phone number in E.164 format
      responses:
        '200':
          description: OTP generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid phone number format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/auth/verify-otp:
    post:
      summary: Verify OTP and get JWT
      description: Verify OTP and return JWT token for authentication
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
                - code
              properties:
                phone:
                  type: string
                  pattern: '^\+[1-9]\d{1,14}$'
                  example: "+15551234567"
                  description: Phone number in E.164 format
                code:
                  type: string
                  example: "123456"
                  description: 6-digit OTP code
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          user:
                            $ref: '#/components/schemas/User'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/{id}:
    get:
      summary: Get user by ID
      description: Retrieve a single user by their ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          example: "123e4567-e89b-12d3-a456-426614174000"
          description: User ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users:
    get:
      summary: List users
      description: Retrieve a paginated list of users with optional search
      tags:
        - Users
      parameters:
        - in: query
          name: phone
          schema:
            type: string
          example: "+1555"
          description: Filter by exact phone number match
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
          description: Number of items per page
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/verify-otp
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          example: "success"
        error:
          type: string
          example: "invalid_phone_format"
        data:
          type: object
          description: Response data (varies by endpoint)
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        phone:
          type: string
          pattern: '^\+[1-9]\d{1,14}$'
          example: "+15551234567"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    UserPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20